# frontend/openapi/openapi.json
{
  "components": {
    "schemas": {
      "Body_uploadSpeech": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_uploadSpeech",
        "type": "object"
      },
      "CategoryFrequency": {
        "properties": {
          "category": {
            "$ref": "#/components/schemas/MistakeCategory"
          },
          "occurrences": {
            "minimum": 0.0,
            "title": "Occurrences",
            "type": "integer"
          },
          "opportunities": {
            "minimum": 0.0,
            "title": "Opportunities",
            "type": "integer"
          }
        },
        "required": [
          "category",
          "occurrences",
          "opportunities"
        ],
        "title": "CategoryFrequency",
        "type": "object"
      },
      "ChangePasswordRequest": {
        "properties": {
          "current_password": {
            "maxLength": 128,
            "minLength": 8,
            "title": "Current Password",
            "type": "string"
          },
          "new_password": {
            "maxLength": 128,
            "minLength": 8,
            "title": "New Password",
            "type": "string"
          }
        },
        "required": [
          "current_password",
          "new_password"
        ],
        "title": "ChangePasswordRequest",
        "type": "object"
      },
      "DateRange": {
        "properties": {
          "end": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End"
          },
          "start": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start"
          }
        },
        "title": "DateRange",
        "type": "object"
      },
      "DetectedMistake": {
        "properties": {
          "category": {
            "$ref": "#/components/schemas/MistakeCategory"
          },
          "correction": {
            "title": "Correction",
            "type": "string"
          },
          "explanation": {
            "title": "Explanation",
            "type": "string"
          },
          "original_text": {
            "title": "Original Text",
            "type": "string"
          }
        },
        "required": [
          "category",
          "original_text",
          "correction",
          "explanation"
        ],
        "title": "DetectedMistake",
        "type": "object"
      },
      "DistributionRequest": {
        "properties": {
          "date_range": {
            "$ref": "#/components/schemas/DateRange"
          }
        },
        "required": [
          "date_range"
        ],
        "title": "DistributionRequest",
        "type": "object"
      },
      "DistributionResponse": {
        "properties": {
          "mistake_frequencies": {
            "items": {
              "$ref": "#/components/schemas/CategoryFrequency"
            },
            "title": "Mistake Frequencies",
            "type": "array"
          },
          "total_speeches": {
            "title": "Total Speeches",
            "type": "integer"
          }
        },
        "required": [
          "total_speeches",
          "mistake_frequencies"
        ],
        "title": "DistributionResponse",
        "type": "object"
      },
      "ErrorBody": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail",
          "code"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorCode": {
        "enum": [
          "UNEXPECTED",
          "INVALID_CREDENTIALS",
          "INVALID_CURRENT_PASSWORD",
          "INVALID_TOKEN",
          "ALREADY_REGISTERED",
          "SPEECH_NOT_FOUND",
          "INVALID_AUDIO",
          "VALIDATION_ERROR"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
      "LoginRequest": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 128,
            "minLength": 8,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
      "LoginResponse": {
        "properties": {
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id"
        ],
        "title": "LoginResponse",
        "type": "object"
      },
      "MistakeCategory": {
        "enum": [
          "subject_verb_agreement",
          "verb_tense",
          "article_usage",
          "preposition_usage",
          "word_order",
          "plurality"
        ],
        "title": "MistakeCategory",
        "type": "string"
      },
      "RegisterRequest": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 128,
            "minLength": 8,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "RegisterRequest",
        "type": "object"
      },
      "RegisterResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "created_at"
        ],
        "title": "RegisterResponse",
        "type": "object"
      },
      "SpeechAnalysis": {
        "properties": {
          "feedback": {
            "title": "Feedback",
            "type": "string"
          },
          "frequencies": {
            "items": {
              "$ref": "#/components/schemas/CategoryFrequency"
            },
            "title": "Frequencies",
            "type": "array"
          },
          "mistakes": {
            "items": {
              "$ref": "#/components/schemas/DetectedMistake"
            },
            "title": "Mistakes",
            "type": "array"
          }
        },
        "required": [
          "frequencies",
          "mistakes",
          "feedback"
        ],
        "title": "SpeechAnalysis",
        "type": "object"
      },
      "SpeechListResponse": {
        "properties": {
          "speeches": {
            "items": {
              "$ref": "#/components/schemas/SpeechResponse"
            },
            "title": "Speeches",
            "type": "array"
          }
        },
        "required": [
          "speeches"
        ],
        "title": "SpeechListResponse",
        "type": "object"
      },
      "SpeechResponse": {
        "properties": {
          "analysis": {
            "$ref": "#/components/schemas/SpeechAnalysis"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "transcript": {
            "title": "Transcript",
            "type": "string"
          }
        },
        "required": [
          "id",
          "transcript",
          "analysis",
          "created_at"
        ],
        "title": "SpeechResponse",
        "type": "object"
      },
      "TimeSeriesPoint": {
        "properties": {
          "occurrences": {
            "title": "Occurrences",
            "type": "integer"
          },
          "opportunities": {
            "title": "Opportunities",
            "type": "integer"
          },
          "time": {
            "format": "date-time",
            "title": "Time",
            "type": "string"
          }
        },
        "required": [
          "occurrences",
          "opportunities",
          "time"
        ],
        "title": "TimeSeriesPoint",
        "type": "object"
      },
      "TimeSeriesRequest": {
        "properties": {
          "date_range": {
            "$ref": "#/components/schemas/DateRange"
          },
          "mistake_category": {
            "$ref": "#/components/schemas/MistakeCategory"
          }
        },
        "required": [
          "date_range",
          "mistake_category"
        ],
        "title": "TimeSeriesRequest",
        "type": "object"
      },
      "TimeSeriesResponse": {
        "properties": {
          "points": {
            "items": {
              "$ref": "#/components/schemas/TimeSeriesPoint"
            },
            "title": "Points",
            "type": "array"
          }
        },
        "required": [
          "points"
        ],
        "title": "TimeSeriesResponse",
        "type": "object"
      }
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "in": "cookie",
        "name": "session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "English Learning Progress Tracker API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/account": {
      "delete": {
        "operationId": "deleteAccount",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Authentication required or invalid"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unexpected server error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Delete Account",
        "tags": [
          "account"
        ]
      }
    },
    "/api/v1/account/password": {
      "patch": {
        "operationId": "changePassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Authentication required or invalid"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unexpected server error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Change Password",
        "tags": [
          "account"
        ]
      }
    },
    "/api/v1/analytics/distribution": {
      "post": {
        "operationId": "getDistribution",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DistributionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DistributionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Authentication required or invalid"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unexpected server error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Get Distribution",
        "tags": [
          "analytics"
        ]
      }
    },
    "/api/v1/analytics/time-series": {
      "post": {
        "operationId": "getTimeSeries",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TimeSeriesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimeSeriesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Authentication required or invalid"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unexpected server error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Get Time Series",
        "tags": [
          "analytics"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Authentication required or invalid"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unexpected server error"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "logout",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unexpected server error"
          }
        },
        "summary": "Logout",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Resource conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unexpected server error"
          }
        },
        "summary": "Register",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/session": {
      "get": {
        "operationId": "getSession",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Authentication required or invalid"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unexpected server error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Get Session",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/speeches": {
      "get": {
        "operationId": "listSpeeches",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpeechListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Authentication required or invalid"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unexpected server error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "List Speeches",
        "tags": [
          "speeches"
        ]
      },
      "post": {
        "operationId": "uploadSpeech",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_uploadSpeech"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpeechResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Authentication required or invalid"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unexpected server error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Upload Speech",
        "tags": [
          "speeches"
        ]
      }
    },
    "/api/v1/speeches/{speech_id}": {
      "delete": {
        "operationId": "deleteSpeech",
        "parameters": [
          {
            "in": "path",
            "name": "speech_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Speech Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Authentication required or invalid"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unexpected server error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Delete Speech",
        "tags": [
          "speeches"
        ]
      }
    }
  }
}